Review comments pt. 1
